attic/io: Rename StreamHasher to HashReader
Makes all naming consistent.
