+++ attic/src/io/mod.rs
@@ -1,29 +1,19 @@
 //! Stream utilities.
 
+mod hash_reader;
+
 use std::collections::VecDeque;
 use std::future::Future;
 use std::marker::Unpin;
 use std::pin::Pin;
-use std::sync::Arc;
-use std::task::{Context, Poll};
 
 use async_stream::try_stream;
 use bytes::{Bytes, BytesMut};
-use digest::{Digest, Output as DigestOutput};
 use futures::stream::{BoxStream, Stream, StreamExt};
-use tokio::io::{AsyncRead, AsyncReadExt, ReadBuf};
-use tokio::sync::OnceCell;
+use tokio::io::{AsyncRead, AsyncReadExt};
 use tokio::task::spawn;
 
-/// Stream filter that hashes the bytes that have been read.
-///
-/// The hash is finalized when EOF is reached.
-pub struct StreamHasher<R: AsyncRead + Unpin, D: Digest + Unpin> {
-    inner: R,
-    digest: Option<D>,
-    bytes_read: usize,
-    finalized: Arc<OnceCell<(DigestOutput<D>, usize)>>,
-}
+pub use hash_reader::HashReader;
 
 /// Merge chunks lazily into a continuous stream.
 ///
@@ -98,60 +88,6 @@ where
     Box::pin(s)
 }
 
-impl<R: AsyncRead + Unpin, D: Digest + Unpin> StreamHasher<R, D> {
-    pub fn new(inner: R, digest: D) -> (Self, Arc<OnceCell<(DigestOutput<D>, usize)>>) {
-        let finalized = Arc::new(OnceCell::new());
-
-        (
-            Self {
-                inner,
-                digest: Some(digest),
-                bytes_read: 0,
-                finalized: finalized.clone(),
-            },
-            finalized,
-        )
-    }
-}
-
-impl<R: AsyncRead + Unpin, D: Digest + Unpin> AsyncRead for StreamHasher<R, D> {
-    fn poll_read(
-        mut self: Pin<&mut Self>,
-        cx: &mut Context<'_>,
-        buf: &mut ReadBuf<'_>,
-    ) -> Poll<tokio::io::Result<()>> {
-        let old_filled = buf.filled().len();
-        let r = Pin::new(&mut self.inner).poll_read(cx, buf);
-        let read_len = buf.filled().len() - old_filled;
-
-        match r {
-            Poll::Ready(Ok(())) => {
-                if read_len == 0 {
-                    // EOF
-                    if let Some(digest) = self.digest.take() {
-                        self.finalized
-                            .set((digest.finalize(), self.bytes_read))
-                            .expect("Hash has already been finalized");
-                    }
-                } else {
-                    // Read something
-                    let digest = self.digest.as_mut().expect("Stream has data after EOF");
-
-                    let filled = buf.filled();
-                    digest.update(&filled[filled.len() - read_len..]);
-                    self.bytes_read += read_len;
-                }
-            }
-            Poll::Ready(Err(_)) => {
-                assert!(read_len == 0);
-            }
-            Poll::Pending => {}
-        }
-
-        r
-    }
-}
-
 /// Greedily reads from a stream to fill a buffer.
 pub async fn read_chunk_async<S: AsyncRead + Unpin + Send>(
     stream: &mut S,
@@ -175,47 +111,6 @@ mod tests {
     use async_stream::stream;
     use bytes::{BufMut, BytesMut};
     use futures::future;
-    use tokio::io::AsyncReadExt;
-
-    #[tokio::test]
-    async fn test_stream_hasher() {
-        let expected = b"hello world";
-        let expected_sha256 =
-            hex::decode("b94d27b9934d3e08a52e52d7da7dabfac484efe37a5380ee9088f7ace2efcde9")
-                .unwrap();
-
-        let (mut read, finalized) = StreamHasher::new(expected.as_slice(), sha2::Sha256::new());
-        assert!(finalized.get().is_none());
-
-        // force multiple reads
-        let mut buf = vec![0u8; 100];
-        let mut bytes_read = 0;
-        bytes_read += read
-            .read(&mut buf[bytes_read..bytes_read + 5])
-            .await
-            .unwrap();
-        bytes_read += read
-            .read(&mut buf[bytes_read..bytes_read + 5])
-            .await
-            .unwrap();
-        bytes_read += read
-            .read(&mut buf[bytes_read..bytes_read + 5])
-            .await
-            .unwrap();
-        bytes_read += read
-            .read(&mut buf[bytes_read..bytes_read + 5])
-            .await
-            .unwrap();
-
-        assert_eq!(expected.len(), bytes_read);
-        assert_eq!(expected, &buf[..bytes_read]);
-
-        let (hash, count) = finalized.get().expect("Hash wasn't finalized");
-
-        assert_eq!(expected_sha256.as_slice(), hash.as_slice());
-        assert_eq!(expected.len(), *count);
-        eprintln!("finalized = {:x?}", finalized);
-    }
 
     #[tokio::test]
     async fn test_merge_chunks() {
